Move hooks to relevant component.
